+++ ProyectosNetbeansWS/TouresBalonProductosWS/src/main/java/aes/pica/touresbalon/touresbalonproductosws/entidades/productos/TarifaHospedaje.java
@@ -14,7 +14,7 @@ public class TarifaHospedaje  implements java.io.Serializable {
 
      private int idHospedaje;
      private String nombreHospedaje;
-     private BigDecimal precio;
+     private double precio;
      private Set productos = new HashSet(0);
 
     public TarifaHospedaje() {
@@ -24,7 +24,7 @@ public class TarifaHospedaje  implements java.io.Serializable {
     public TarifaHospedaje(int idHospedaje) {
         this.idHospedaje = idHospedaje;
     }
-    public TarifaHospedaje(int idHospedaje, String nombreHospedaje, BigDecimal precio, Set productos) {
+    public TarifaHospedaje(int idHospedaje, String nombreHospedaje, double precio, Set productos) {
        this.idHospedaje = idHospedaje;
        this.nombreHospedaje = nombreHospedaje;
        this.precio = precio;
@@ -45,11 +45,11 @@ public class TarifaHospedaje  implements java.io.Serializable {
     public void setNombreHospedaje(String nombreHospedaje) {
         this.nombreHospedaje = nombreHospedaje;
     }
-    public BigDecimal getPrecio() {
+    public double getPrecio() {
         return this.precio;
     }
     
-    public void setPrecio(BigDecimal precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
     public Set getProductos() {
